ready for testing

+++ Src/SSSF/src/NetworkStats/NetworkStats.h
@@ -18,7 +18,7 @@ public:
     struct HealthBasics
     {
         int64_t lastMessageTime = 0;
-        uint32_t lastSequenceNumber = 0;
+        int64_t lastSequenceNumber = 0;
     };
 
     struct HealthCore
